split core distance to serial and parallel
